requirements.txt
asgiref==3.6.0
black==23.1.0
click==8.1.3
colorama==0.4.6
Django==4.1.6
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
pi==0.1.2
platformdirs==3.0.0
PyJWT==1.7.1
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
